LPS-8521 SafeFileNameHookWrapper not taking into consideration old files that may not have the same name format
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@48947 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/documentlibrary/util/SafeFileNameHookWrapper.java
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.service.ServiceContext;
+import com.liferay.portlet.documentlibrary.model.DLFileEntryConstants;
 
 import java.io.File;
 import java.io.InputStream;
@@ -39,8 +40,17 @@ public class SafeFileNameHookWrapper implements Hook {
 	public void addDirectory(long companyId, long repositoryId, String dirName)
 		throws PortalException, SystemException {
 
-		_hook.addDirectory(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(dirName));
+		String safeDirName = FileUtil.encodeSafeFileName(dirName);
+
+		if (!safeDirName.equals(dirName)) {
+			try {
+				_hook.move(dirName, safeDirName);
+			}
+			catch (Exception e) {
+			}
+		}
+
+		_hook.addDirectory(companyId, repositoryId, safeDirName);
 	}
 
 	public void addFile(
@@ -49,10 +59,15 @@ public class SafeFileNameHookWrapper implements Hook {
 			Date modifiedDate, ServiceContext serviceContext, byte[] bytes)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+
 		_hook.addFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName), fileEntryId, properties,
-			modifiedDate, serviceContext, bytes);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			fileEntryId, properties, modifiedDate, serviceContext, bytes);
 	}
 
 	public void addFile(
@@ -61,10 +76,15 @@ public class SafeFileNameHookWrapper implements Hook {
 			Date modifiedDate, ServiceContext serviceContext, File file)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+
 		_hook.addFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName), fileEntryId, properties,
-			modifiedDate, serviceContext, file);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			fileEntryId, properties, modifiedDate, serviceContext, file);
 	}
 
 	public void addFile(
@@ -73,10 +93,15 @@ public class SafeFileNameHookWrapper implements Hook {
 			Date modifiedDate, ServiceContext serviceContext, InputStream is)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+
 		_hook.addFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName), fileEntryId,
-			properties, modifiedDate, serviceContext, is);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			fileEntryId, properties, modifiedDate, serviceContext, is);
 	}
 
 	public void checkRoot(long companyId) throws SystemException {
@@ -87,9 +112,20 @@ public class SafeFileNameHookWrapper implements Hook {
 			long companyId, String portletId, long repositoryId, String dirName)
 		throws PortalException, SystemException {
 
-		_hook.deleteDirectory(
-			companyId, portletId, repositoryId,
-			FileUtil.encodeSafeFileName(dirName));
+		String safeDirName = FileUtil.encodeSafeFileName(dirName);
+
+		if (!safeDirName.equals(dirName)) {
+			try {
+				_hook.deleteDirectory(
+					companyId, portletId, repositoryId, dirName);
+
+				return;
+			}
+			catch (Exception e) {
+			}
+		}
+
+		_hook.deleteDirectory(companyId, portletId, repositoryId, safeDirName);
 	}
 
 	public void deleteFile(
@@ -97,9 +133,18 @@ public class SafeFileNameHookWrapper implements Hook {
 			String fileName)
 		throws PortalException, SystemException {
 
-		_hook.deleteFile(
-			companyId, portletId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName));
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName,
+				DLFileEntryConstants.DEFAULT_VERSION)) {
+
+			_hook.deleteFile(companyId, portletId, repositoryId, fileName);
+
+			return;
+		}
+
+		_hook.deleteFile(companyId, portletId, repositoryId, safeFileName);
 	}
 
 	public void deleteFile(
@@ -107,16 +152,34 @@ public class SafeFileNameHookWrapper implements Hook {
 			String fileName, String versionNumber)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName, versionNumber)) {
+
+			_hook.deleteFile(
+				companyId, portletId, repositoryId, fileName, versionNumber);
+
+			return;
+		}
+
 		_hook.deleteFile(
-			companyId, portletId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName), versionNumber);
+			companyId, portletId, repositoryId, safeFileName, versionNumber);
 	}
 
 	public byte[] getFile(long companyId, long repositoryId, String fileName)
 		throws PortalException, SystemException {
 
-		return _hook.getFile(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(fileName));
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName,
+				DLFileEntryConstants.DEFAULT_VERSION)) {
+
+			return _hook.getFile(companyId, repositoryId, fileName);
+		}
+
+		return _hook.getFile(companyId, repositoryId, safeFileName);
 	}
 
 	public byte[] getFile(
@@ -124,17 +187,33 @@ public class SafeFileNameHookWrapper implements Hook {
 			String versionNumber)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName, versionNumber)) {
+
+			return _hook.getFile(
+				companyId, repositoryId, fileName, versionNumber);
+		}
+
 		return _hook.getFile(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(fileName),
-			versionNumber);
+			companyId, repositoryId, safeFileName, versionNumber);
 	}
 
 	public InputStream getFileAsStream(
 			long companyId, long repositoryId, String fileName)
 		throws PortalException, SystemException {
 
-		return _hook.getFileAsStream(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(fileName));
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName,
+				DLFileEntryConstants.DEFAULT_VERSION)) {
+
+			return _hook.getFileAsStream(companyId, repositoryId, fileName);
+		}
+
+		return _hook.getFileAsStream(companyId, repositoryId, safeFileName);
 	}
 
 	public InputStream getFileAsStream(
@@ -142,17 +221,35 @@ public class SafeFileNameHookWrapper implements Hook {
 			String versionNumber)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName, versionNumber)) {
+
+			return _hook.getFileAsStream(
+				companyId, repositoryId, fileName, versionNumber);
+		}
+
 		return _hook.getFileAsStream(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(fileName),
-			versionNumber);
+			companyId, repositoryId, safeFileName, versionNumber);
 	}
 
 	public String[] getFileNames(
 			long companyId, long repositoryId, String dirName)
 		throws PortalException, SystemException {
 
+		String safeDirName = FileUtil.encodeSafeFileName(dirName);
+
+		if (!safeDirName.equals(dirName)) {
+			try {
+				_hook.move(dirName, safeDirName);
+			}
+			catch (Exception e) {
+			}
+		}
+
 		String[] fileNames = _hook.getFileNames(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(dirName));
+			companyId, repositoryId, safeDirName);
 
 		String[] decodedFileNames = new String[fileNames.length];
 
@@ -167,8 +264,16 @@ public class SafeFileNameHookWrapper implements Hook {
 			long companyId, long repositoryId, String fileName)
 		throws PortalException, SystemException {
 
-		return _hook.getFileSize(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(fileName));
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) && _hook.hasFile(
+				companyId, repositoryId, fileName,
+				DLFileEntryConstants.DEFAULT_VERSION)) {
+
+			return _hook.getFileSize(companyId, repositoryId, fileName);
+		}
+
+		return _hook.getFileSize(companyId, repositoryId, safeFileName);
 	}
 
 	public boolean hasFile(
@@ -176,9 +281,16 @@ public class SafeFileNameHookWrapper implements Hook {
 			String versionNumber)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		if (!safeFileName.equals(fileName) &&
+			_hook.hasFile(companyId, repositoryId, fileName, versionNumber)) {
+
+			return true;
+		}
+
 		return _hook.hasFile(
-			companyId, repositoryId, FileUtil.encodeSafeFileName(fileName),
-			versionNumber);
+			companyId, repositoryId, safeFileName, versionNumber);
 	}
 
 	public void move(String srcDir, String destDir) throws SystemException {
@@ -194,9 +306,15 @@ public class SafeFileNameHookWrapper implements Hook {
 			long newRepositoryId, String fileName, long fileEntryId)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+
 		_hook.updateFile(
 			companyId, portletId, groupId, repositoryId, newRepositoryId,
-			FileUtil.encodeSafeFileName(fileName), fileEntryId);
+			safeFileName, fileEntryId);
 	}
 
 	public void updateFile(
@@ -206,11 +324,20 @@ public class SafeFileNameHookWrapper implements Hook {
 			ServiceContext serviceContext, byte[] bytes)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+		String safeSourceFileName = FileUtil.encodeSafeFileName(sourceFileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, sourceFileName,
+			safeSourceFileName);
+
 		_hook.updateFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName), versionNumber,
-			FileUtil.encodeSafeFileName(sourceFileName), fileEntryId,
-			properties, modifiedDate, serviceContext, bytes);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			versionNumber, safeSourceFileName, fileEntryId, properties,
+			modifiedDate, serviceContext, bytes);
 	}
 
 	public void updateFile(
@@ -220,11 +347,20 @@ public class SafeFileNameHookWrapper implements Hook {
 			ServiceContext serviceContext, File file)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+		String safeSourceFileName = FileUtil.encodeSafeFileName(sourceFileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, sourceFileName,
+			safeSourceFileName);
+
 		_hook.updateFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName),
-			versionNumber, FileUtil.encodeSafeFileName(sourceFileName),
-			fileEntryId, properties, modifiedDate, serviceContext, file);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			versionNumber, safeSourceFileName, fileEntryId, properties,
+			modifiedDate, serviceContext, file);
 	}
 
 	public void updateFile(
@@ -234,11 +370,20 @@ public class SafeFileNameHookWrapper implements Hook {
 			ServiceContext serviceContext, InputStream is)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+		String safeSourceFileName = FileUtil.encodeSafeFileName(sourceFileName);
+
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, fileName,
+			safeFileName);
+		renameUnsafeFile(
+			companyId, portletId, groupId, repositoryId, sourceFileName,
+			safeSourceFileName);
+
 		_hook.updateFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName), versionNumber,
-			FileUtil.encodeSafeFileName(sourceFileName), fileEntryId,
-			properties, modifiedDate, serviceContext, is);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			versionNumber, safeSourceFileName, fileEntryId, properties,
+			modifiedDate, serviceContext, is);
 	}
 
 	public void updateFile(
@@ -246,10 +391,38 @@ public class SafeFileNameHookWrapper implements Hook {
 			String fileName, String newFileName, boolean reindex)
 		throws PortalException, SystemException {
 
+		String safeFileName = FileUtil.encodeSafeFileName(fileName);
+		String safeNewFileName = FileUtil.encodeSafeFileName(newFileName);
+
+		if (!safeFileName.equals(fileName)) {
+			if (_hook.hasFile(
+					companyId, repositoryId, fileName,
+					DLFileEntryConstants.DEFAULT_VERSION)) {
+
+				safeFileName = fileName;
+			}
+		}
+
 		_hook.updateFile(
-			companyId, portletId, groupId, repositoryId,
-			FileUtil.encodeSafeFileName(fileName),
-			FileUtil.encodeSafeFileName(newFileName), reindex);
+			companyId, portletId, groupId, repositoryId, safeFileName,
+			safeNewFileName, reindex);
+	}
+
+	protected void renameUnsafeFile(
+			long companyId, String portletId, long groupId, long repositoryId,
+			String fileName, String safeFileName)
+		throws PortalException, SystemException {
+
+		if (!safeFileName.equals(fileName)) {
+			if (_hook.hasFile(
+					companyId, repositoryId, fileName,
+					DLFileEntryConstants.DEFAULT_VERSION)) {
+
+				_hook.updateFile(
+					companyId, portletId, groupId, repositoryId, fileName,
+					safeFileName, true);
+			}
+		}
 	}
 
 	private Hook _hook;
